swaybar: separate input code to new file
