+++ include/swaybar/i3bar.h
@@ -1,7 +1,7 @@
 #ifndef _SWAYBAR_I3BAR_H
 #define _SWAYBAR_I3BAR_H
 
-#include "bar.h"
+#include "input.h"
 #include "status_line.h"
 
 struct i3bar_block {
@@ -28,7 +28,5 @@ void i3bar_block_unref(struct i3bar_block *block);
 bool i3bar_handle_readable(struct status_line *status);
 enum hotspot_event_handling i3bar_block_send_click(struct status_line *status,
 		struct i3bar_block *block, int x, int y, enum x11_button button);
-enum x11_button wl_button_to_x11_button(uint32_t button);
-enum x11_button wl_axis_to_x11_button(uint32_t axis, wl_fixed_t value);
 
 #endif
