+++ swaybar/meson.build
@@ -3,6 +3,7 @@ executable(
 		'bar.c',
 		'config.c',
 		'i3bar.c',
+		'input.c',
 		'ipc.c',
 		'main.c',
 		'render.c',
